Revert "Changing to doubles to floats to for native m1 mac support"
This reverts commit cdb05afb28ab4c8bb8329a05704c56543eb5664c.

+++ src/nanopyx/opencl/_cl_mandelbrot_benchmark.py
@@ -37,7 +37,7 @@ def _cl_mandelbrot(
         None,
         im_mandelbrot.data,
         np.int32(max_iter),
-        np.float32(divergence),
+        np.float64(divergence),
     )
     queue.finish()
